requirements.txt
ultralytics>=8.1.0
opencv-python>=4.8
numpy>=1.24
pyyaml
matplotlib
ffmpeg
deep_sort_realtime
torch>=2.0
torchvision>=0.17
pandas
streamlit==1.40.0          # 或 1.39.x 以下
streamlit-drawable-canvas==0.9.3
tqdm
matplotlib